feat(basket): add updateRestaurant method and throw Invalid request if necessary

+++ services/basket/src/handlers/basket/reset.ts
@@ -9,6 +9,10 @@ export const Reset = async (
 ) => {
     try {
         const {id} = request;
+        if (!id) {
+            throw new Error("Invalid request");
+        }
+
         await client.del(`user:${id}`);
         callback(null, {});
     } catch (error) {
